Fix partially initialized module
- Moved 'CreateThread' from globals to core module, only used by core anyways
